+++ test/orchard/resource_test.clj
@@ -1,6 +1,7 @@
 (ns orchard.resource-test
   (:require
-   [clojure.test :refer [deftest testing is]]
+   [clojure.string :as str]
+   [clojure.test :refer [deftest is testing]]
    [orchard.resource :as resource]))
 
 (deftest resource-path-tuple-test
@@ -8,5 +9,7 @@
 
 (deftest project-resources-test
   (testing "get the correct resources for the orchard project"
-    (is (= "see-also.edn" (-> (resource/project-resources) first :relpath)))
-    (is (= java.net.URL (-> (resource/project-resources) first :url class)))))
+    (let [resources (->> (resource/project-resources)
+                         (remove #(str/ends-with? (.getAbsolutePath (:root %)) "test-resources")))]
+      (is (= "see-also.edn" (-> resources first :relpath)))
+      (is (= java.net.URL (-> resources first :url class))))))